Complete figure 14
Modify run_sim.tcl to accept more parameters. Add a function that
parses the output of the nam.out file to get the throughput of a
simulation. Use link rates of 100Mbps for now and increase the
link delay for figure 14 so that improvement can be observed as the
value of K increases. Need to find what is wrong and 1Gbs and 10Gbps
simulations do not work for any of the figures.

Signed-off-by: Kostis Kaffes <kkaffes@stanford.edu>